requirements.txt
requests
python-dotenv
aiogram
aiohttp
asyncio
deep_translator
bleach